add updater function for BasicInfo
